watsonTests/proj/requirements.txt
# Framework requirements
watson-framework

# Project requirements
pytest
pytest-cov
